Add tests and reformat
