+++ serde-tests/test.rs
@@ -1,0 +1,1 @@
+include!(concat!(env!("OUT_DIR"), "/test.rs"));
