+++ src/ordered.rs
@@ -1,16 +1,13 @@
 use chrono::{DateTime, UTC};
-use bson::{Array,Bson,Document};
-use super::spec::BinarySubtype;
-use super::oid::ObjectId;
 use std::collections::BTreeMap;
-use std::error;
-use std::fmt;
+use std::{error, fmt, slice};
 use std::fmt::{Debug, Display, Error, Formatter};
 use std::iter::{FromIterator, Map};
 use std::vec::IntoIter;
-use std::slice;
-use serde::ser::{Serialize, Serializer};
-use serde::ser::impls::MapIteratorVisitor;
+
+use bson::{Array,Bson,Document};
+use oid::ObjectId;
+use spec::BinarySubtype;
 
 /// Error to indicate that either a value was empty or it contained an unexpected
 /// type, for use with the direct getters.
@@ -365,15 +362,6 @@ impl OrderedDocument {
             self.keys.remove(position.unwrap());
         }
         self.document.remove(key)
-    }
-}
-
-impl Serialize for OrderedDocument {
-    #[inline]
-    fn serialize<S>(&self, serializer: &mut S) -> Result<(), S::Error>
-        where S: Serializer,
-    {
-        serializer.visit_map(MapIteratorVisitor::new(self.iter(), Some(self.len())))
     }
 }
